fix(invites): allow empty accept body and drop email-domain org enrich
Inject {} for missing accept-invite POST bodies so clients without a payload
no longer hit missing_payload, and limit org enrichment to auth-service
metadata instead of inferring from the email domain. Consolidate related tests
into http_test.go, committee_service_test.go, and organization_helpers_test.go.
